# kmgen,swift: [swift] big performance optimization for Swift 5.7