That change considerably reduces the dependency on karmem/swift
package, consequently reducing Swift's generics code.

The generated code will use "inline" code. It will also use the
new `loadUnaligned`.
```
name               old time/op    new time/op    delta
DecodeSumVec3-8       112ms ±12%       9ms ± 3%  -91.95%  (p=0.008 n=5+5)
DecodeObjectAPI-8    88.1ms ± 7%    30.8ms ±32%  -65.08%  (p=0.016 n=4+5)
EncodeObjectAPI-8    67.0ms ±56%     8.2ms ± 0%  -87.72%  (p=0.008 n=5+5)

name               old alloc/op   new alloc/op   delta
DecodeSumVec3-8      2.98kB ± 0%    1.82kB ± 0%     ~     (p=0.079 n=4+5)
DecodeObjectAPI-8     359kB ±16%     108kB ± 3%  -69.85%  (p=0.008 n=5+5)
EncodeObjectAPI-8    2.96kB ± 0%    1.81kB ± 0%  -38.92%  (p=0.008 n=5+5)

name               old allocs/op  new allocs/op  delta
DecodeSumVec3-8        33.0 ± 0%      32.0 ± 0%   -3.03%  (p=0.008 n=5+5)
DecodeObjectAPI-8      33.0 ± 0%      32.0 ± 0%   -3.03%  (p=0.008 n=5+5)
EncodeObjectAPI-8      31.0 ± 0%      30.0 ± 0%   -3.23%  (p=0.008 n=5+5)
```

Signed-off-by: Inkeliz <inkeliz@inkeliz.com>